+++ src/main/java/com/accsaber/backend/model/dto/request/map/ImportMapFromLeaderboardIdsRequest.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.request.map;
 
-import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
@@ -28,7 +27,7 @@ public class ImportMapFromLeaderboardIdsRequest {
     @NotBlank
     private String characteristic;
 
-    private BigDecimal complexity;
+    private Double complexity;
 
     private UUID batchId;
