+++ src/main/java/com/accsaber/backend/model/dto/response/campaign/BarrierProgressResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.campaign;
 
-import java.math.BigDecimal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -10,7 +9,7 @@ import lombok.Getter;
 public class BarrierProgressResponse {
 
     private CampaignBarrierResponse barrier;
-    private BigDecimal currentValue;
+    private Double currentValue;
     private boolean satisfied;
     private boolean unlocked;
 }
