+++ src/main/java/com/accsaber/backend/model/dto/response/campaign/CampaignNodeScoreEntry.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.campaign;
 
-import java.math.BigDecimal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -12,6 +11,6 @@ public class CampaignNodeScoreEntry {
     private Integer rank;
     private CampaignLeaderboardPlayer player;
     private Integer score;
-    private BigDecimal accuracy;
-    private BigDecimal ap;
+    private Double accuracy;
+    private Double ap;
 }
