+++ src/main/java/com/accsaber/backend/model/dto/response/campaign/CampaignTargetProgressResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.campaign;
 
-import java.math.BigDecimal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -10,6 +9,6 @@ import lombok.Getter;
 public class CampaignTargetProgressResponse {
 
     private CampaignTargetResponse target;
-    private BigDecimal userValue;
+    private Double userValue;
     private boolean met;
 }
