+++ src/main/java/com/accsaber/backend/model/dto/response/campaign/CampaignTargetResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.campaign;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import com.accsaber.backend.model.entity.campaign.CampaignRequirementType;
@@ -14,6 +13,6 @@ public class CampaignTargetResponse {
 
     private UUID id;
     private CampaignRequirementType requirementType;
-    private BigDecimal requirementValue;
-    private BigDecimal requirementValueMax;
+    private Double requirementValue;
+    private Double requirementValueMax;
 }
