+++ src/main/java/com/accsaber/backend/model/dto/response/campaign/CampaignTextResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.campaign;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 import lombok.Builder;
@@ -12,10 +11,10 @@ public class CampaignTextResponse {
 
     private UUID id;
     private String content;
-    private BigDecimal positionX;
-    private BigDecimal positionY;
+    private Double positionX;
+    private Double positionY;
     private String font;
-    private BigDecimal scale;
+    private Double scale;
     private String color;
     private String effects;
 }
