+++ src/main/java/com/accsaber/backend/model/dto/response/item/CrateContentResponse.java
@@ -1,6 +1,5 @@
 package com.accsaber.backend.model.dto.response.item;
 
-import java.math.BigDecimal;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -11,5 +10,5 @@ public class CrateContentResponse {
 
     private ItemResponse rewardItem;
     private Integer dropWeight;
-    private BigDecimal dropChance;
+    private Double dropChance;
 }
